[Spring] Extract LinearSpring class in its own file
