+++ Sofa/Component/SolidMechanics/Spring/CMakeLists.txt
@@ -20,6 +20,7 @@ set(HEADER_FILES
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/JointSpring.inl
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/JointSpringForceField.h
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/JointSpringForceField.inl
+    ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/LinearSpring.h
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/MeshSpringForceField.h
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/MeshSpringForceField.inl
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/PolynomialRestShapeSpringsForceField.h
@@ -61,6 +62,7 @@ set(SOURCE_FILES
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/GearSpringForceField.cpp
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/JointSpring.cpp
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/JointSpringForceField.cpp
+    ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/LinearSpring.cpp
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/MeshSpringForceField.cpp
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/PolynomialRestShapeSpringsForceField.cpp
     ${SOFACOMPONENTSOLIDMECHANICSSPRING_SOURCE_DIR}/PolynomialSpringsForceField.cpp
